小程序属性标签删除条件修改

+++ skyeye-service/src/main/java/com/skyeye/smprogram/service/impl/RmPropertyServiceImpl.java
@@ -79,10 +79,28 @@ public class RmPropertyServiceImpl implements RmPropertyService{
 		Map<String, Object> map = inputObject.getParams();
 		Map<String, Object> bean = rmPropertyDao.queryRmPropertyValueNumById(map);
 		if(bean == null){
-			rmPropertyDao.deleteRmPropertyMationById(map);
-		}else{
-			if(Integer.parseInt(bean.get("propertyValueNum").toString()) == 0){//该样式属性下有值
+			Map<String, Object> useThisBean = rmPropertyDao.queryUseRmPropertyNumById(map);
+			if(useThisBean == null){
 				rmPropertyDao.deleteRmPropertyMationById(map);
+			}else{
+				if(Integer.parseInt(useThisBean.get("usePropertyNum").toString()) == 0){//该样式属性没有被使用
+					rmPropertyDao.deleteRmPropertyMationById(map);
+				}else{
+					outputObject.setreturnMessage("该样式属性正在使用中，无法删除。");
+				}
+			}
+		}else{
+			if(Integer.parseInt(bean.get("propertyValueNum").toString()) == 0){//该样式属性下没有值
+				Map<String, Object> useThisBean = rmPropertyDao.queryUseRmPropertyNumById(map);
+				if(useThisBean == null){
+					rmPropertyDao.deleteRmPropertyMationById(map);
+				}else{
+					if(Integer.parseInt(useThisBean.get("usePropertyNum").toString()) == 0){//该样式属性没有被使用
+						rmPropertyDao.deleteRmPropertyMationById(map);
+					}else{
+						outputObject.setreturnMessage("该样式属性正在使用中，无法删除。");
+					}
+				}
 			}else{
 				outputObject.setreturnMessage("该样式属性下存在值，无法删除。");
 			}
